+++ src/test/converter2/issues/gh3009.ts
@@ -1,0 +1,7 @@
+/**
+ * [link with base path](src/test/converter2/issues/gh3009.ts)
+ * [link direct](gh3009.ts)
+ *
+ * [asset link](CHANGELOG.md)
+ */
+export const x = 1;